Move `ThemeFamily` and `ThemeVariant` into crate root
